Modify a-intnam.ads & handler.S LPC43 files manually
The SVD file for LPC43 is missing definitions for some of the interrupts
therefore svd2ada generated file a-intnam.ads is also missing name and
ID for those interrupts. File handler.S is consistent with vector table
but comments are not, so fixed those as well. Manually modifying auto
generated files is not the correct way to fix it. The correct way is to
fix the SVD file.