Fix lpc43 clocking issue
UART had framing issues and it was determined that it had to do with
clocking issues. Error was with MSEL value. MSEL value is PLL1's feeedback
divider division ratio. For 204MHz clock MSEL value should be 17, but
register starts at 1 with 0x00 value, so to set MSEL to 17 register
value should be set to MSEL-1, so register value should have been set to
16 not 17.